Refactor DriverSettings, removed redundant classes
