--- src/main/java/aquality/appium/mobile/configuration/driversettings/IOSSettings.java
@@ -1,16 +1,0 @@
-package aquality.appium.mobile.configuration.driversettings;
-
-import aquality.appium.mobile.application.PlatformName;
-import aquality.selenium.core.utilities.ISettingsFile;
-
-public class IOSSettings extends DriverSettings{
-
-    public IOSSettings(ISettingsFile settingsFile){
-        super(settingsFile);
-    }
-
-    @Override
-    public PlatformName getPlatformName() {
-        return PlatformName.IOS;
-    }
-}
